Move auth tests to unit tests
format auth tests
