+++ requirements-dev.txt
@@ -2,5 +2,6 @@ flake8>=3.7
 black>=18.9b0
 pytest>=5.3
 pytest-asyncio>=0.10.0
+pytest-mock>=2.0.0
 git+https://github.com/dask/dask@master
 git+https://github.com/dask/distributed@master
